requirements.txt
scapy
python-nmap
bcrypt 
sqlalchemy
requests
bleak
beautifulsoup4
